requirements.txt
av==10.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==3.16.0
faster-whisper==0.6.0
filelock==3.12.2
flatbuffers==23.5.26
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-speech==2.20.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauth2client==4.1.3
onnxruntime==1.15.1
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pyparsing==3.1.0
pyreadline3==3.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.12
tiktoken==0.3.3
tokenizers==0.13.3
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.65.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
